API-650: Rework Create client integration test
